requirements.txt
streamlit
langchain-core
langchain-openai
langchain-community
openai
faiss-cpu
PyMuPDF
python-dotenv
